docs: reference appropriate image
There is no way (ATM) to inspect container from the container
itself, so we can not detect what image is the source of the
actual container.

The workaround is to "dynamically" specify, during container
build, the real namespace/name:tag as is going to be used in
dockerhub repository.

* Makefile (docker_main_tag_check): Set to true to stop build if
the $docker_main_tag variable is not specified.
* cont-lib: Move forward.
* project.py: Bump minor version.
* share/cont-docs/postgresql.txt.tpl: Use docker_main_tag macro
instead of static string.
* NEWS: Document.